edge_glass_modular/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "edge-glass-modular"
version = "1.0.0"
description = "Modular multimodal alignment and instruction tuning framework"
authors = [{name = "Edge Glass Team"}]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchaudio>=2.0.0",
    "transformers>=4.35.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "peft>=0.6.0",
    "bitsandbytes>=0.41.0",
    "sentence-transformers>=2.2.0",
    "pillow>=10.0.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "einops>=0.7.0",
    "wandb>=0.15.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100
